test(tools): split cancellation cleanup coverage
